Move DNS Cleanup task to the cleanup_vm role
This change makes it easy to customize Generic Cloud images. Before
this change, the user should add their package installation or network
related tasks before this task to avoid any name resolution errors
which is was reported by the community members that could be hard find
the reason for the issue.

Signed-off-by: Elkhan Mammadli <elkhan.mammadli@protonmail.com>