Fix and enhancements for AMIs
- Fixes for AlmaLinux OS 8
On AlmaLinux OS 8, The UUID of root partition should be set in
GRUB environment block to make the boot process work on UEFI.

- Set serial as a primary console
We tried to make kernel options compatible with
RHEL Amazon Machine Images. However, the order of the console devices
is not consistent between the builds and major versions:
"console=ttyS0,115200n8 console=tty0" vs
"console=tty0 console=ttyS0,115200n8".
Since the last defined console is the primary, which means it used
also by systemd to print init logs in addition the kernel logs.

Let's set "console=ttyS0,115200n8" as a primary,
as the last console to print the whole output (kernel and init)
to EC2 serial console.

This should aid troubleshooting/debugging experience of boot process.

Signed-off-by: Elkhan Mammadli <elkhan.mammadli@protonmail.com>