feat(s390x): add sources and documentation
We cannot use Packer to build s390x images because of the
zipl (the zSeries Initial Program Loader), the bootloader
implementation of s390x architecture. Packer cannot interrupt the boot
process to inject the URL of the kickstart from its HTTP server to
process the installation.

Since the working mechanism of the zipl is the similar to BIOS,
which installs the boot loader code on the initial sectors of the disk.
We need an alternative tool which uses Virtualization with a disk to
write the boot loader code instead of a tool with chroot-like approach.

We decided to choose Oz because it has been used by CentOS Stream Build
Service to build cloud images including the s390x images which are
functional. While borrowing the idea of the tool to build those images,
we did not deter ourselves to borrow the kickstart files as a starting
point and modify for our needs.

Let's publish the kickstart files and TDL templatez of Oz with
instructions about the building Generic Cloud images of AlmaLinux OS 9
and 8 for s390x.

Signed-off-by: Elkhan Mammadli <elkhan.mammadli@protonmail.com>