fix(kickstart): create required partitions
When installing an BIOS system into GPT partition table, biosboot
partition is needed as a first partition on the disk. Adding reqpart to
the kickstart file of BIOS versions of the kickstart files, tells
Anaconda Installer to automatically create partitions required by the
hardware platform.

This fixes error on Anaconda Installer due the partitioning.

Signed-off-by: Elkhan Mammadli <elkhan.mammadli@protonmail.com>