feat(9): overhaul installed packages
Overhauling the list of installed packages on all cloud images and
storing the packages lists provides some benefits:

- Controllable: This helps to make the sizes of the cloud images
as light as possible. When we get a build large in size,
it will be easier to track down the reason.
- Predictability: Users can consult to the packages lists to organize
their automation with the required packages.
It also makes easier to start a discussion to add or
remove a package from an image.

Signed-off-by: Elkhan Mammadli <elkhan.mammadli@protonmail.com>