chore(cleanup_vm): moving the removal of persistent net rules (#202)
- This couple of Ansible tasks make sure there is not persistent net
rules left behind on /etc/udev/rules.d (e.g. 70-persistent-net.rules).
Since nature of this operation is cleaning up the system for the image
creation. It is better to move this operation from the gencloud_guest
Ansible role to the cleanup_vm.

- Use the --oldinstallonly command-line argument of dnf remove command
to remove old kernel and packages. Which specifically serves to this
purpose.

Signed-off-by: Elkhan Mammadli <elkhan.mammadli@protonmail.com>