chore(8): optimize unification of bootloader configurations (#224)
The current method of configuration of GRUB2 environment block:

- Remove /boot/grub2/grubenv which is a symbolic link to
/boot/efi/EFI/almalinux/grubenv.
- Remove /boot/efi/EFI/almalinux/grubenv which created during
installation by the Anaconda installer.
- Use an unified_boot_kernel_opts Ansible variable for the images which
has own optimized kernel command-line parameters like Azure and OCI.
- Generate a new GRUB2 environment block using the
unified_boot_kernel_opts Ansible variable from the scratch.

The new method of configuration of GRUB2 environment block:

- Copy /boot/efi/EFI/almalinux/grubenv file by overwriting the
/boot/grub2/grubenv symbolic link.

The new approach is simpler and needs less code and makes all the
kernel command-line parameters are defined on the kickstart
configurations permanent. It will also contribute the next patch
which is intended for the 9.5 and 8.10 updates of OCI images.

Signed-off-by: Elkhan Mammadli <elkhan.mammadli@protonmail.com>