fix(10-ami): fix package installation and naming (#247)
ansible.builtin.dnf uses the host's RPM repositories to install
dependencies instead of system on the installroot parameter.
The source of this bug is might be the DNF Python API which is a
requirement and used by of the ansible.builtin.dnf.

Which is results with:
- Error on different major versions e.g. Build AlmaLinux OS 10 AMI on
AlmaLinux OS 9.
- Installation of wrong packages into the installroot. For instance,
when build an AlmaLinux OS 10 on top of the AlmaLinux OS Kitten 10. It
installs the AlmaLinux OS Kitten 10 versions of the packages instead of
the AlmaLinux OS 10.

Using dnf itself is fixes these two cases.

- Fixed AMI naming by adding the full version of the AlmaLinux OS 10
- Fixed AMI Description and version to have full version identifier.

Signed-off-by: Elkhan Mammadli <elkhan.mammadli@protonmail.com>